Address review feedback
- Consolidate ws-labs.ts into labs.ts file
- Sort preview features by localized integration name
- Add haStyle and clean up h1/h2 styling
- Apply spacing variable suggestions in dialogs
- Use ha-space-* variables for consistent spacing
